# users-microservice/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/sign-up":{"post":{"operationId":"UsersMicroserviceController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":""}},"tags":["UsersMicroservice"]}},"/api/v1/login":{"post":{"operationId":"UsersMicroserviceController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["UsersMicroservice"]}},"/api/v1/refresh-token":{"post":{"operationId":"UsersMicroserviceController_reAuth","parameters":[],"responses":{"201":{"description":""}},"tags":["UsersMicroservice"]}},"/api/v1/me":{"get":{"operationId":"UsersMicroserviceController_me","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["UsersMicroservice"]}},"/api/v1/device-sessions":{"get":{"operationId":"DeviceSessionsController_getDeviceSessions","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["device-sessions"]}},"/api/v1/device-sessions/logout":{"post":{"operationId":"DeviceSessionsController_logout","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["device-sessions"]}},"/api/v1/friends/requests/pending":{"get":{"operationId":"FriendController_getPendingRequests","parameters":[],"responses":{"200":{"description":""}},"tags":["Friend"]}},"/api/v1/friends/suggestions":{"get":{"operationId":"FriendController_getSuggestedFriends","parameters":[],"responses":{"200":{"description":""}},"tags":["Friend"]}},"/api/v1/friends/requests/send/{receiverId}":{"post":{"operationId":"FriendController_sendFriendRequest","parameters":[{"name":"receiverId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Friend"]}},"/api/v1/friends/requests/{requestId}/accept":{"post":{"operationId":"FriendController_acceptFriendRequest","parameters":[{"name":"requestId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Friend"]}},"/api/v1/friends/requests/{requestId}/reject":{"post":{"operationId":"FriendController_rejectFriendRequest","parameters":[{"name":"requestId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Friend"]}},"/api/v1/user-profiles":{"post":{"operationId":"UserProfileController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserProfileDto"}}}},"responses":{"201":{"description":""}},"tags":["UserProfile"]},"get":{"operationId":"UserProfileController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["UserProfile"]}},"/api/v1/user-profiles/{id}":{"get":{"operationId":"UserProfileController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["UserProfile"]},"patch":{"operationId":"UserProfileController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserProfileDto"}}}},"responses":{"200":{"description":""}},"tags":["UserProfile"]},"delete":{"operationId":"UserProfileController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["UserProfile"]}}},"info":{"title":"User","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignUpDto":{"type":"object","properties":{"email":{"type":"string","example":"mail@example.com"},"password":{"type":"string"},"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"},"role":{"type":"string","example":"USER"},"status":{"type":"string","example":"ACTIVE"},"params":{"type":"object","example":"{}"}},"required":["email","password","firstName","lastName"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"mail@example.com"},"password":{"type":"string","example":"Linh@123"}},"required":["email","password"]},"LogoutDto":{"type":"object","properties":{"sessionId":{"type":"string"}},"required":["sessionId"]},"CreateUserProfileDto":{"type":"object","properties":{}},"UpdateUserProfileDto":{"type":"object","properties":{}}}}}